+++ backend/handlers/auth.js
@@ -2,27 +2,34 @@ const jwt = require("jsonwebtoken"),
   User = require("../models/user"),
   CustomError = require("../middleware/error").CustomError;
 
-const sendJSONResponse = (req, res, next) => {
+const sendNewToken = (req, res, next) => {
   // get the user's username and id
-  const { username, id } = req.user;
+  const { username, id, extensions } = req.user;
   const userData = {
     username,
     id
   };
 
-  // if something goes wrong and req.user
-  // has the password and other info, we don't want to
-  // send it back with the jwt by accident
-
   // create new token and send it back to the user
   const token = jwt.sign(userData, process.env.JWT_SECRET);
-  res.json({ token, user: userData });
+  res.json({ token, user: { ...userData, extensions } });
   // if there is a next function run it
   // so we don't need to specify the
   // next function when we don't need it
   if (next !== undefined) {
     next();
   }
+};
+
+const sendUserJSON = (req, res) => {
+  User.findById(req.user.id).then(user => {
+    const userJSON = {
+      username: user.username,
+      id: user._id,
+      extensions: user.extensions
+    };
+    res.json({ user: userJSON });
+  });
 };
 
 const registerHandler = (req, res) => {
@@ -49,6 +56,7 @@ const registerHandler = (req, res) => {
 };
 
 module.exports = {
-  sendJSONResponse,
-  registerHandler
+  sendNewToken,
+  registerHandler,
+  sendUserJSON
 };