+++ backend/routes/extensions.js
@@ -1,0 +1,11 @@
+const router = require("express").Router(),
+  verifyToken = require("../middleware/auth").verifyToken,
+  extensionHandlers = require("../handlers/extensions");
+
+router.use(verifyToken);
+
+router.post("/", extensionHandlers.addExtension);
+
+router.delete("/:extensionName", extensionHandlers.removeExtension);
+
+module.exports = router;