+++ webclient/src/components/LandingPageSections/Pomodoro/Pomodoro.js
@@ -1,22 +1,26 @@
 import React from "react";
-import clockIcon from "../../../assets/clock.svg";
+import activityImage from "../../../assets/landing-activity.svg";
 
 const pomodoro = () => {
   return (
     <section className="Section Section-Pomodoro">
-      <div className="Section-Wrapper Container">
-        <div className="Subsection">
+      <div className="Container row">
+        <div className="Subsection col m8">
           <h2 className="Subsection-Header">
             Focus on your work. <br />
             Get rid of distractions.
           </h2>
           <p className="Subsection-Paragraph">
-            With 25 minute sessions dedicated to work and 5 minute breaks, your
-            work becomes your main focus again.
+            Stay on top of your time with 25 minute sessions dedicated to work
+            and 5 minute breaks in between.
           </p>
         </div>
-        <div className="Subsection">
-          <img className="Subsection-Image" src={clockIcon} alt="Clock Icon" />
+        <div className="Subsection col m4">
+          <img
+            className="Subsection-Image"
+            src={activityImage}
+            alt="Clock Icon"
+          />
         </div>
       </div>
     </section>