+++ webclient/src/store/reducers/auth.js
@@ -1,4 +1,5 @@
 import * as actionTypes from "../actions/actionTypes";
+import _ from "lodash";
 
 const initialState = {
   // undefined = not checked
@@ -11,6 +12,7 @@ const initialState = {
 };
 
 export default (state = initialState, action) => {
+  let newState = _.cloneDeep(state);
   switch (action.type) {
     case actionTypes.AUTH_START:
       return { ...state, loading: true, error: null };
@@ -26,6 +28,17 @@ export default (state = initialState, action) => {
       return { ...state, loading: false, error: action.error };
     case actionTypes.AUTH_LOGOUT:
       return { ...state, token: null, user: null };
+    case actionTypes.ADD_EXTENSION: {
+      newState.user.extensions.push(action.extensionName);
+      return newState;
+    }
+    case actionTypes.REMOVE_EXTENSION: {
+      const filtered = newState.user.extensions.filter(
+        e => e !== action.extensionName
+      );
+      newState.user.extensions = filtered;
+      return newState;
+    }
     default:
       return state;
   }